pyproject.toml
[project]
name = "agent"
version = "4.0.0"
description = "Add your description here"
authors = [
    { name = "Marc Fabian Mezger", email = "marc.mezger@gmail.com" }
]

dependencies = [
    "langchain-cohere>=0.3.1",
    "loguru>=0.7.2",
    "omegaconf>=2.3.0",
    "tiktoken>=0.8.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.4",
    "openai>=1.53.0",
    "langchain-qdrant>=0.1.4",
    "uvicorn>=0.32.0",
    "pypdfium2>=4.30.0",
    "ultra-simple-config>=0.1.0",
    "langchain-openai>=0.2.3",
    "langchain-community>=0.3.3",
    "langgraph>=0.2.39",
    "arize-phoenix>=5.5.2",
    "fastembed>=0.4.1",
    "nltk>=3.9.1",
    "lingua-language-detector>=2.0.2",
    "weave>=0.51.17",
    "openinference-instrumentation-langchain>=0.1.28",
    "arize-phoenix-otel>=0.6.1",
    "langchain-ollama>=0.2.0",
    "diagrams>=0.23.4",
    "pymupdf4llm>=0.0.17",
]

readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "yappi>=1.6.0",
    "mypy>=1.11.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
